added comb2.c from proj/hw as this is very useful
I just realized that the algorithm for combination does not have to be somehow similar to permutation.
No, it has to be exactly the same!
Since P(n, r) = n! / (n - r)!  and C(n, r) = n! / (n - r)! * r!, it should have been obvious from the beginning that the combination of n taken r at a time is actually the permutation of n taken n at a time but the objects are dissected into two groups of similar objects, namely, 1 and 0.
The fundamental principle states that if P is a permutation of n objects taken n at a time but q objects are similar and r objects are similar and s objects are similar and so on, then P = n! / q! * r! * s! ...
So, C(n, r) = P(n, n) / (n - r)! * r!
In simpler terms, the objects in a combination are places in a permutation and the places in a combination are objects in a permutation.